// src/models/BookmarkModels.ts
// Generated by https://quicktype.io

import { HashSet } from "../HashSet";

export interface BookmarkFolderModel {
    id: string;
    name: string;
    bgColor: string;
    bookmarks: HashSet<BookmarkModel>;
}

export interface BookmarkModel {
    id: string;
    name: string;
    url: string;
    bgColor: string;
}